Oulines changes to UI organization
