+++ broker/alert_acquisition/schema/ztf_3.2.json
