+++ broker/data_upload/__init__.py
@@ -8,6 +8,5 @@ information on the pricing models associated with various data upload services.
 
 from ._ingest_alerts import _batch_ingest
 from ._ingest_alerts import get_schema
-from ._ingest_alerts import save_to_avro
 from ._ingest_alerts import _stream_ingest
 from ._ingest_alerts import upload_to_bucket
